[step2] fix: 각각의 로또 당첨 번호 로직 수정
