+++ src/main/java/lotto/domain/Lottos.java
@@ -6,14 +6,8 @@ import java.util.List;
 public class Lottos {
     private List<Lotto> lottoList;
 
-    public Lottos() {}
-
     public Lottos(List<Lotto> lottoList) {
         this.lottoList = lottoList;
-    }
-
-    public int getNumberOfLotto() {
-        return lottoList.size();
     }
 
     public HashMap<WinningAmount, Integer> countAllWinning(List<Integer> winningNumber) {
