+++ src/main/java/lotto/domain/WinningLottoManager.java
@@ -4,21 +4,20 @@ import java.util.HashMap;
 import java.util.List;
 
 public class WinningLottoManager {
-    List<Integer> winningNumber;
-    HashMap<WinningAmount, Integer> winningLottos = new HashMap<>();
+    private final List<Integer> WINNING_NUMBER;
+    private HashMap<WinningAmount, Integer> winningLottos = new HashMap<>();
 
     public WinningLottoManager(List<Integer> winningNumber) {
-        this.winningNumber = winningNumber;
+        this.WINNING_NUMBER = winningNumber;
         for (WinningAmount winningAmount : WinningAmount.values()) {
             winningLottos.put(winningAmount, 0);
         }
     }
 
     public void countWinning(Lotto lotto) {
-        Integer countMatch = lotto.countWinningNumber(winningNumber);
+        Integer countMatch = lotto.countWinningNumber(WINNING_NUMBER);
         WinningAmount winningAmount = WinningAmount.findWinningAmountByMatchCount(countMatch);
         mergeIfMoreThanThreeMatch(winningAmount);
-
     }
 
     private void mergeIfMoreThanThreeMatch(WinningAmount winningAmount) {
